Create Path.Posix and Path.Windows, redistribute code
