+++ .gitignore
@@ -8,3 +8,4 @@ TAGS
 tags
 *.tag
 .stack-work/
+*.swp
