+++ src/Path/Posix.hs
@@ -1,0 +1,4 @@
+{-# LANGUAGE CPP #-}
+#define PLATFORM_NAME   Posix
+#define IS_WINDOWS      False
+#include "Include.hs"
