+++ src/Path/Windows.hs
@@ -1,0 +1,4 @@
+{-# LANGUAGE CPP #-}
+#define PLATFORM_NAME   Windows
+#define IS_WINDOWS      True
+#include "Include.hs"
